fix(common): correct coordinates type

+++ app/common.c
@@ -3,7 +3,7 @@
 static uint8_t cell_length;
 static point_t field_start;
 
-void cm_set_cell_length(uint8_t length)
+void cm_set_cell_length(uint16_t length)
 {
     cell_length = length;
 }
@@ -13,7 +13,7 @@ uint8_t cm_get_cell_length(void)
     return cell_length;
 }
 
-void cm_set_field_start(uint8_t x, uint8_t y)
+void cm_set_field_start(uint16_t x, uint16_t y)
 {
     field_start.x = x;
     field_start.y = y;
