+++ app/common.h
@@ -5,14 +5,14 @@
 
 typedef struct
 {
-    uint8_t x;
-    uint8_t y;
+    uint16_t x;
+    uint16_t y;
 } point_t;
 
 //TODO doc
-void cm_set_cell_length(uint8_t length);
+void cm_set_cell_length(uint16_t length);
 uint8_t cm_get_cell_length(void);
-void cm_set_field_start(uint8_t x, uint8_t y);
+void cm_set_field_start(uint16_t x, uint16_t y);
 point_t cm_get_field_start(void);
 
 #endif /*COMMON_H_*/
